✨ #25 feat: security, handler 추가, 수정
