--- src/main/java/Capstone/Davinchi_Server/global/config/WebSecurityConfig.java
@@ -1,82 +1,0 @@
-package Capstone.Davinchi_Server.global.config;
-
-import Capstone.Davinchi_Server.global.exception.AuthenticationEntryPointHandler;
-import Capstone.Davinchi_Server.global.jwt.JwtAuthenticationFilter;
-import Capstone.Davinchi_Server.global.jwt.JwtTokenUtil;
-import Capstone.Davinchi_Server.oauth2.CustomOAuth2UserService;
-import Capstone.Davinchi_Server.oauth2.HttpCookieOAuth2AuthorizationRequestRepository;
-import Capstone.Davinchi_Server.oauth2.OAuth2AuthenticationFailureHandler;
-import Capstone.Davinchi_Server.oauth2.OAuth2AuthenticationSuccessHandler;
-import lombok.AllArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.servlet.handler.HandlerMappingIntrospector;
-
-@Configuration
-@AllArgsConstructor
-@EnableWebSecurity
-public class WebSecurityConfig {
-
-    private final JwtTokenUtil jwtTokenUtil;
-    private final AuthenticationEntryPointHandler authenticationEntryPointHandler;
-    private final CustomOAuth2UserService customOAuth2UserService;
-    private final OAuth2AuthenticationSuccessHandler oAuth2AuthenticationSuccessHandler;
-    private final OAuth2AuthenticationFailureHandler oAuth2AuthenticationFailureHandler;
-
-    @Bean
-    public HttpCookieOAuth2AuthorizationRequestRepository cookieOAuth2AuthorizationRequestRepository() {
-        return new HttpCookieOAuth2AuthorizationRequestRepository();
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http, HandlerMappingIntrospector introspector) throws Exception {
-        http
-                .httpBasic(AbstractHttpConfigurer::disable)
-                .csrf(AbstractHttpConfigurer::disable)
-                .cors(AbstractHttpConfigurer::disable)
-                .sessionManagement((sessionManagent) -> sessionManagent
-                        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                )
-                .authorizeHttpRequests((authorizeHttpRequests) -> authorizeHttpRequests
-                        .requestMatchers("/jwt/**", "/oauth2/**", "/login/**").permitAll()
-                        .requestMatchers( "/","/swagger-ui/**", "/v3/api-docs/**").permitAll()
-                        .requestMatchers( "/actuator/**").permitAll()
-                        .anyRequest().authenticated()
-                )
-                .exceptionHandling((exceptionConfig) ->
-                        exceptionConfig.authenticationEntryPoint(authenticationEntryPointHandler))
-                // JwtAuthenticationFilter를 UsernamePasswordAuthenticationFilter 전에 넣는다
-                .addFilterBefore(new JwtAuthenticationFilter(jwtTokenUtil), UsernamePasswordAuthenticationFilter.class)
-
-                //oauth2
-                .oauth2Login(oauth2Login -> oauth2Login
-                        .authorizationEndpoint(authorizationEndpoint ->
-                                authorizationEndpoint
-                                        //권한 부여 엔드포인트
-                                        .baseUri("/oauth2/authorize")
-                                        //권한 요청 저장
-                                        .authorizationRequestRepository(cookieOAuth2AuthorizationRequestRepository())
-                        )
-                        .redirectionEndpoint(redirectionEndpoint ->
-                                redirectionEndpoint
-                                        //
-                                        .baseUri("/login/oauth2/code/**")
-                        )
-                        .userInfoEndpoint(userInfoEndpoint ->
-                                userInfoEndpoint
-                                        .userService(customOAuth2UserService)
-                        )
-                        .successHandler(oAuth2AuthenticationSuccessHandler)
-//                        .failureHandler(oAuth2AuthenticationFailureHandler)
-                );
-
-        return http.build();
-    }
-
-}
